+++ lux/_config/__init__.py
@@ -4,4 +4,5 @@ from lux._config.config import (
     remove_action,
     actions,
     update_actions,
+    config,
 )
