+++ lux/_config/config.py
@@ -4,6 +4,7 @@ For more resources, see https://github.com/pandas-dev/pandas/blob/master/pandas/
 '''
 from collections import namedtuple
 from typing import Any, Callable, Dict, Iterable, List, Optional
+import warnings
 
 RegisteredOption = namedtuple("RegisteredOption", "name action display_condition args")
 
@@ -142,4 +143,30 @@ def is_callable(obj) -> bool:
 	if not callable(obj):
 		raise ValueError("Value must be a callable")
 	return True
-	
+
+class Config:
+
+	def __init__(self):
+		self._default_display = "pandas"
+
+	@property
+	def default_display(self):
+		return self._default_display
+
+	@default_display.setter
+	def default_display(self, type:str) -> None:
+		"""
+		Set the widget display to show Pandas by default or Lux by default
+		Parameters
+		----------
+		type : str
+			Default display type, can take either the string `lux` or `pandas` (regardless of capitalization)
+		"""
+		if (type.lower()=="lux"):
+			self._default_display = "lux"
+		elif (type.lower()=="pandas"):
+			self._default_display = "pandas"
+		else: 
+			warnings.warn("Unsupported display type. Default display option should either be `lux` or `pandas`.",stacklevel=2)
+
+config = Config()
